# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Model Computer Vision API",
    "description": "\nMulti-model computer vision API serving image classification, object detection,\nand batch inference.\n\n## Authentication\n\nExchange an API key for a bearer token at `POST /api/v1/auth/token`, then send\n`Authorization: Bearer <token>` on every request.\n\n## Rate limits\n\nApplied per user tier and reported in `X-RateLimit-Limit` and\n`X-RateLimit-Remaining` on every response. A rejected request returns 429 with\n`Retry-After`.\n\n| Tier | Requests / minute |\n| --- | --- |\n| free | 10 |\n| pro | 120 |\n| enterprise | 1200 |\n\n## Correlation\n\nEvery response carries `X-Correlation-ID`. Supply your own to have it adopted;\nquote it when reporting a problem.\n",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/token": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Exchange an API key for an access token",
        "description": "Exchange a long-lived API key for a short-lived bearer token.\n\nThe token carries the caller's tier, which determines their rate limit.\nShort lifetimes mean a leaked token stops working on its own, without\nneeding the underlying key to be rotated.",
        "operationId": "issue_token_api_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unknown API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Body of an error response."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope returned for every error."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "minLength": 8,
            "title": "Api Key",
            "description": "Issued API key."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "api_key"
        ],
        "title": "TokenRequest",
        "description": "Exchange an API key for a short-lived access token."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "Token lifetime in seconds."
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "tier"
        ],
        "title": "TokenResponse",
        "description": "An issued access token."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}